pyproject.toml
[tool.poetry]
name = "vocoder"
version = "0.1.0"
description = ""
authors = ["ricky <rickycontact9@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "*"
torch = "*"
sounddevice = "*"
webrtcvad = "*"
tqdm = "*"
lark-parser = "*"
loguru = "*"
transformers = "*"
aioconsole = "^0.5.0"
graphviz = "^0.20.1"

[tool.poetry.dev-dependencies]
pytest = "*"
ipython = "*"
black = "*"
isort = "*"
autoflake = "*"
pytest-asyncio = "*"
pytest-mock = "*"
pytest-timeout = "^2.1.0"
invoke = "^1.7.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 88
